+++ lib/tests/test_git.rs
@@ -114,6 +114,23 @@ fn get_git_repo(repo: &Arc<ReadonlyRepo>) -> git2::Repository {
     get_git_backend(repo).open_git_repo().unwrap()
 }
 
+fn get_git_executable_path(subprocess: bool) -> Option<PathBuf> {
+    subprocess
+        .then(|| {
+            std::env::var("TEST_GIT_PATH")
+                .map(|x| x.into())
+                .or_else(|e| {
+                    if matches!(e, std::env::VarError::NotPresent) {
+                        Ok("git".into())
+                    } else {
+                        Err(e)
+                    }
+                })
+        })
+        .transpose()
+        .unwrap()
+}
+
 #[test]
 fn test_import_refs() {
     let git_settings = GitSettings {
@@ -2524,15 +2541,18 @@ fn test_init() {
     assert!(!repo.view().heads().contains(&jj_id(&initial_git_commit)));
 }
 
-#[test]
-fn test_fetch_empty_repo() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_fetch_empty_repo(subprocess: bool) {
     let test_data = GitRepoData::create();
     let git_settings = GitSettings::default();
+    let git_executable_path = get_git_executable_path(subprocess);
 
     let mut tx = test_data.repo.start_transaction();
     let stats = git::fetch(
         tx.repo_mut(),
         &test_data.git_repo,
+        git_executable_path.as_deref(),
         "origin",
         &[StringPattern::everything()],
         git::RemoteCallbacks::default(),
@@ -2547,19 +2567,22 @@ fn test_fetch_empty_repo() {
     assert_eq!(tx.repo_mut().view().bookmarks().count(), 0);
 }
 
-#[test]
-fn test_fetch_initial_commit_head_is_not_set() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_fetch_initial_commit_head_is_not_set(subprocess: bool) {
     let test_data = GitRepoData::create();
     let git_settings = GitSettings {
         auto_local_bookmark: true,
         ..Default::default()
     };
     let initial_git_commit = empty_git_commit(&test_data.origin_repo, "refs/heads/main", &[]);
+    let git_executable_path = get_git_executable_path(subprocess);
 
     let mut tx = test_data.repo.start_transaction();
     let stats = git::fetch(
         tx.repo_mut(),
         &test_data.git_repo,
+        git_executable_path.as_deref(),
         "origin",
         &[StringPattern::everything()],
         git::RemoteCallbacks::default(),
@@ -2598,8 +2621,9 @@ fn test_fetch_initial_commit_head_is_not_set() {
     );
 }
 
-#[test]
-fn test_fetch_initial_commit_head_is_set() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_fetch_initial_commit_head_is_set(subprocess: bool) {
     let test_data = GitRepoData::create();
     let git_settings = GitSettings {
         auto_local_bookmark: true,
@@ -2616,11 +2640,13 @@ fn test_fetch_initial_commit_head_is_set() {
         .origin_repo
         .reference("refs/tags/v1.0", new_git_commit.id(), false, "")
         .unwrap();
+    let git_executable_path = get_git_executable_path(subprocess);
 
     let mut tx = test_data.repo.start_transaction();
     let stats = git::fetch(
         tx.repo_mut(),
         &test_data.git_repo,
+        git_executable_path.as_deref(),
         "origin",
         &[StringPattern::everything()],
         git::RemoteCallbacks::default(),
@@ -2633,19 +2659,22 @@ fn test_fetch_initial_commit_head_is_set() {
     assert!(stats.import_stats.abandoned_commits.is_empty());
 }
 
-#[test]
-fn test_fetch_success() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_fetch_success(subprocess: bool) {
     let mut test_data = GitRepoData::create();
     let git_settings = GitSettings {
         auto_local_bookmark: true,
         ..Default::default()
     };
     let initial_git_commit = empty_git_commit(&test_data.origin_repo, "refs/heads/main", &[]);
+    let git_executable_path = get_git_executable_path(subprocess);
 
     let mut tx = test_data.repo.start_transaction();
     git::fetch(
         tx.repo_mut(),
         &test_data.git_repo,
+        git_executable_path.as_deref(),
         "origin",
         &[StringPattern::everything()],
         git::RemoteCallbacks::default(),
@@ -2670,6 +2699,7 @@ fn test_fetch_success() {
     let stats = git::fetch(
         tx.repo_mut(),
         &test_data.git_repo,
+        git_executable_path.as_deref(),
         "origin",
         &[StringPattern::everything()],
         git::RemoteCallbacks::default(),
@@ -2715,19 +2745,22 @@ fn test_fetch_success() {
     );
 }
 
-#[test]
-fn test_fetch_prune_deleted_ref() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_fetch_prune_deleted_ref(subprocess: bool) {
     let test_data = GitRepoData::create();
     let git_settings = GitSettings {
         auto_local_bookmark: true,
         ..Default::default()
     };
     let commit = empty_git_commit(&test_data.origin_repo, "refs/heads/main", &[]);
+    let git_executable_path = get_git_executable_path(subprocess);
 
     let mut tx = test_data.repo.start_transaction();
     git::fetch(
         tx.repo_mut(),
         &test_data.git_repo,
+        git_executable_path.as_deref(),
         "origin",
         &[StringPattern::everything()],
         git::RemoteCallbacks::default(),
@@ -2752,6 +2785,7 @@ fn test_fetch_prune_deleted_ref() {
     let stats = git::fetch(
         tx.repo_mut(),
         &test_data.git_repo,
+        git_executable_path.as_deref(),
         "origin",
         &[StringPattern::everything()],
         git::RemoteCallbacks::default(),
@@ -2767,19 +2801,22 @@ fn test_fetch_prune_deleted_ref() {
         .is_absent());
 }
 
-#[test]
-fn test_fetch_no_default_branch() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_fetch_no_default_branch(subprocess: bool) {
     let test_data = GitRepoData::create();
     let git_settings = GitSettings {
         auto_local_bookmark: true,
         ..Default::default()
     };
     let initial_git_commit = empty_git_commit(&test_data.origin_repo, "refs/heads/main", &[]);
+    let git_executable_path = get_git_executable_path(subprocess);
 
     let mut tx = test_data.repo.start_transaction();
     git::fetch(
         tx.repo_mut(),
         &test_data.git_repo,
+        git_executable_path.as_deref(),
         "origin",
         &[StringPattern::everything()],
         git::RemoteCallbacks::default(),
@@ -2804,6 +2841,7 @@ fn test_fetch_no_default_branch() {
     let stats = git::fetch(
         tx.repo_mut(),
         &test_data.git_repo,
+        git_executable_path.as_deref(),
         "origin",
         &[StringPattern::everything()],
         git::RemoteCallbacks::default(),
@@ -2815,17 +2853,20 @@ fn test_fetch_no_default_branch() {
     assert_eq!(stats.default_branch, None);
 }
 
-#[test]
-fn test_fetch_empty_refspecs() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_fetch_empty_refspecs(subprocess: bool) {
     let test_data = GitRepoData::create();
     let git_settings = GitSettings::default();
     empty_git_commit(&test_data.origin_repo, "refs/heads/main", &[]);
+    let git_executable_path = get_git_executable_path(subprocess);
 
     // Base refspecs shouldn't be respected
     let mut tx = test_data.repo.start_transaction();
     git::fetch(
         tx.repo_mut(),
         &test_data.git_repo,
+        git_executable_path.as_deref(),
         "origin",
         &[],
         git::RemoteCallbacks::default(),
@@ -2845,14 +2886,17 @@ fn test_fetch_empty_refspecs() {
         .is_absent());
 }
 
-#[test]
-fn test_fetch_no_such_remote() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_fetch_no_such_remote(subprocess: bool) {
     let test_data = GitRepoData::create();
     let git_settings = GitSettings::default();
     let mut tx = test_data.repo.start_transaction();
+    let git_executable_path = get_git_executable_path(subprocess);
     let result = git::fetch(
         tx.repo_mut(),
         &test_data.git_repo,
+        git_executable_path.as_deref(),
         "invalid-remote",
         &[StringPattern::everything()],
         git::RemoteCallbacks::default(),
@@ -2955,13 +2999,15 @@ fn set_up_push_repos(settings: &UserSettings, temp_dir: &TempDir) -> PushTestSet
     }
 }
 
-#[test]
-fn test_push_bookmarks_success() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_push_bookmarks_success(subprocess: bool) {
     let settings = testutils::user_settings();
     let temp_dir = testutils::new_temp_dir();
     let mut setup = set_up_push_repos(&settings, &temp_dir);
     let clone_repo = get_git_repo(&setup.jj_repo);
     let mut tx = setup.jj_repo.start_transaction();
+    let git_executable_path = get_git_executable_path(subprocess);
 
     let targets = GitBranchPushTargets {
         branch_updates: vec![(
@@ -2975,6 +3021,7 @@ fn test_push_bookmarks_success() {
     let result = git::push_branches(
         tx.repo_mut(),
         &clone_repo,
+        git_executable_path.as_deref(),
         "origin",
         &targets,
         git::RemoteCallbacks::default(),
@@ -3020,13 +3067,15 @@ fn test_push_bookmarks_success() {
     assert!(!tx.repo_mut().has_changes());
 }
 
-#[test]
-fn test_push_bookmarks_deletion() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_push_bookmarks_deletion(subprocess: bool) {
     let settings = testutils::user_settings();
     let temp_dir = testutils::new_temp_dir();
     let mut setup = set_up_push_repos(&settings, &temp_dir);
     let clone_repo = get_git_repo(&setup.jj_repo);
     let mut tx = setup.jj_repo.start_transaction();
+    let git_executable_path = get_git_executable_path(subprocess);
 
     let source_repo = git2::Repository::open(&setup.source_repo_dir).unwrap();
     // Test the setup
@@ -3044,6 +3093,7 @@ fn test_push_bookmarks_deletion() {
     let result = git::push_branches(
         tx.repo_mut(),
         &get_git_repo(&setup.jj_repo),
+        git_executable_path.as_deref(),
         "origin",
         &targets,
         git::RemoteCallbacks::default(),
@@ -3072,13 +3122,15 @@ fn test_push_bookmarks_deletion() {
     assert!(!tx.repo_mut().has_changes());
 }
 
-#[test]
-fn test_push_bookmarks_mixed_deletion_and_addition() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_push_bookmarks_mixed_deletion_and_addition(subprocess: bool) {
     let settings = testutils::user_settings();
     let temp_dir = testutils::new_temp_dir();
     let mut setup = set_up_push_repos(&settings, &temp_dir);
     let clone_repo = get_git_repo(&setup.jj_repo);
     let mut tx = setup.jj_repo.start_transaction();
+    let git_executable_path = get_git_executable_path(subprocess);
 
     let targets = GitBranchPushTargets {
         branch_updates: vec![
@@ -3101,6 +3153,7 @@ fn test_push_bookmarks_mixed_deletion_and_addition() {
     let result = git::push_branches(
         tx.repo_mut(),
         &clone_repo,
+        git_executable_path.as_deref(),
         "origin",
         &targets,
         git::RemoteCallbacks::default(),
@@ -3141,12 +3194,14 @@ fn test_push_bookmarks_mixed_deletion_and_addition() {
     assert!(!tx.repo_mut().has_changes());
 }
 
-#[test]
-fn test_push_bookmarks_not_fast_forward() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_push_bookmarks_not_fast_forward(subprocess: bool) {
     let settings = testutils::user_settings();
     let temp_dir = testutils::new_temp_dir();
     let setup = set_up_push_repos(&settings, &temp_dir);
     let mut tx = setup.jj_repo.start_transaction();
+    let git_executable_path = get_git_executable_path(subprocess);
 
     let targets = GitBranchPushTargets {
         branch_updates: vec![(
@@ -3160,6 +3215,7 @@ fn test_push_bookmarks_not_fast_forward() {
     let result = git::push_branches(
         tx.repo_mut(),
         &get_git_repo(&setup.jj_repo),
+        git_executable_path.as_deref(),
         "origin",
         &targets,
         git::RemoteCallbacks::default(),
@@ -3179,11 +3235,13 @@ fn test_push_bookmarks_not_fast_forward() {
 // may want to add tests for when a bookmark unexpectedly moved backwards or
 // unexpectedly does not exist for bookmark deletion.
 
-#[test]
-fn test_push_updates_unexpectedly_moved_sideways_on_remote() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_push_updates_unexpectedly_moved_sideways_on_remote(subprocess: bool) {
     let settings = testutils::user_settings();
     let temp_dir = testutils::new_temp_dir();
     let setup = set_up_push_repos(&settings, &temp_dir);
+    let git_executable_path = get_git_executable_path(subprocess);
 
     // The main bookmark is actually at `main_commit` on the remote. If we expect
     // it to be at `sideways_commit`, it unexpectedly moved sideways from our
@@ -3205,6 +3263,7 @@ fn test_push_updates_unexpectedly_moved_sideways_on_remote() {
         git::push_updates(
             setup.jj_repo.as_ref(),
             &get_git_repo(&setup.jj_repo),
+            git_executable_path.as_deref(),
             "origin",
             &targets,
             git::RemoteCallbacks::default(),
@@ -3244,11 +3303,13 @@ fn test_push_updates_unexpectedly_moved_sideways_on_remote() {
     );
 }
 
-#[test]
-fn test_push_updates_unexpectedly_moved_forward_on_remote() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_push_updates_unexpectedly_moved_forward_on_remote(subprocess: bool) {
     let settings = testutils::user_settings();
     let temp_dir = testutils::new_temp_dir();
     let setup = set_up_push_repos(&settings, &temp_dir);
+    let git_executable_path = get_git_executable_path(subprocess);
 
     // The main bookmark is actually at `main_commit` on the remote. If we
     // expected it to be at `parent_of_commit`, it unexpectedly moved forward
@@ -3272,6 +3333,7 @@ fn test_push_updates_unexpectedly_moved_forward_on_remote() {
         git::push_updates(
             setup.jj_repo.as_ref(),
             &get_git_repo(&setup.jj_repo),
+            git_executable_path.as_deref(),
             "origin",
             &targets,
             git::RemoteCallbacks::default(),
@@ -3298,19 +3360,29 @@ fn test_push_updates_unexpectedly_moved_forward_on_remote() {
         Err(GitPushError::RefInUnexpectedLocation(_))
     );
 
-    // Moving the bookmark *forwards* is OK, as an exception matching our bookmark
-    // conflict resolution rules
-    assert_matches!(
-        attempt_push_expecting_parent(Some(setup.child_of_main_commit.id().clone())),
-        Ok(())
-    );
+    if subprocess {
+        // git is strict about honouring the expected location on --force-with-lease
+        assert_matches!(
+            attempt_push_expecting_parent(Some(setup.child_of_main_commit.id().clone())),
+            Err(GitPushError::RefInUnexpectedLocation(_))
+        );
+    } else {
+        // Moving the bookmark *forwards* is OK, as an exception matching our bookmark
+        // conflict resolution rules
+        assert_matches!(
+            attempt_push_expecting_parent(Some(setup.child_of_main_commit.id().clone())),
+            Ok(())
+        );
+    }
 }
 
-#[test]
-fn test_push_updates_unexpectedly_exists_on_remote() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_push_updates_unexpectedly_exists_on_remote(subprocess: bool) {
     let settings = testutils::user_settings();
     let temp_dir = testutils::new_temp_dir();
     let setup = set_up_push_repos(&settings, &temp_dir);
+    let git_executable_path = get_git_executable_path(subprocess);
 
     // The main bookmark is actually at `main_commit` on the remote. In this test,
     // we expect it to not exist on the remote at all.
@@ -3330,6 +3402,7 @@ fn test_push_updates_unexpectedly_exists_on_remote() {
         git::push_updates(
             setup.jj_repo.as_ref(),
             &get_git_repo(&setup.jj_repo),
+            git_executable_path.as_deref(),
             "origin",
             &targets,
             git::RemoteCallbacks::default(),
@@ -3341,22 +3414,34 @@ fn test_push_updates_unexpectedly_exists_on_remote() {
         Err(GitPushError::RefInUnexpectedLocation(_))
     );
 
-    // We *can* move the bookmark forward even if we didn't expect it to exist
-    assert_matches!(
-        attempt_push_expecting_absence(Some(setup.child_of_main_commit.id().clone())),
-        Ok(())
-    );
+    if subprocess {
+        // Git is strict with enforcing the expected location
+        assert_matches!(
+            attempt_push_expecting_absence(Some(setup.child_of_main_commit.id().clone())),
+            Err(GitPushError::RefInUnexpectedLocation(_))
+        );
+    } else {
+        // In git2: We *can* move the bookmark forward even if we didn't expect it to
+        // exist
+        assert_matches!(
+            attempt_push_expecting_absence(Some(setup.child_of_main_commit.id().clone())),
+            Ok(())
+        );
+    }
 }
 
-#[test]
-fn test_push_updates_success() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_push_updates_success(subprocess: bool) {
     let settings = testutils::user_settings();
     let temp_dir = testutils::new_temp_dir();
     let setup = set_up_push_repos(&settings, &temp_dir);
+    let git_executable_path = get_git_executable_path(subprocess);
     let clone_repo = get_git_repo(&setup.jj_repo);
     let result = git::push_updates(
         setup.jj_repo.as_ref(),
         &clone_repo,
+        git_executable_path.as_deref(),
         "origin",
         &[GitRefUpdate {
             qualified_name: "refs/heads/main".to_string(),
@@ -3386,14 +3471,17 @@ fn test_push_updates_success() {
     assert_eq!(new_target, Some(new_oid));
 }
 
-#[test]
-fn test_push_updates_no_such_remote() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_push_updates_no_such_remote(subprocess: bool) {
     let settings = testutils::user_settings();
     let temp_dir = testutils::new_temp_dir();
     let setup = set_up_push_repos(&settings, &temp_dir);
+    let git_executable_path = get_git_executable_path(subprocess);
     let result = git::push_updates(
         setup.jj_repo.as_ref(),
         &get_git_repo(&setup.jj_repo),
+        git_executable_path.as_deref(),
         "invalid-remote",
         &[GitRefUpdate {
             qualified_name: "refs/heads/main".to_string(),
@@ -3405,14 +3493,17 @@ fn test_push_updates_no_such_remote() {
     assert!(matches!(result, Err(GitPushError::NoSuchRemote(_))));
 }
 
-#[test]
-fn test_push_updates_invalid_remote() {
+#[test_case(false; "use git2 for remote calls")]
+#[test_case(true; "spawn a git subprocess for remote calls")]
+fn test_push_updates_invalid_remote(subprocess: bool) {
     let settings = testutils::user_settings();
     let temp_dir = testutils::new_temp_dir();
     let setup = set_up_push_repos(&settings, &temp_dir);
+    let git_executable_path = get_git_executable_path(subprocess);
     let result = git::push_updates(
         setup.jj_repo.as_ref(),
         &get_git_repo(&setup.jj_repo),
+        git_executable_path.as_deref(),
         "http://invalid-remote",
         &[GitRefUpdate {
             qualified_name: "refs/heads/main".to_string(),